Revert "Move logger interface from usecase to interfaces"
