+++ app/interfaces/post_controller.go
@@ -11,11 +11,11 @@ import (
 // A PostController is a controller for a post.
 type PostController struct {
 	PostInteractor usecase.PostInteractor
-	Logger         Logger
+	Logger         usecase.Logger
 }
 
 // NewPostController creates a PostController.
-func NewPostController(connMySQL *sql.DB, connRedis *redis.Client, logger Logger) *PostController {
+func NewPostController(connMySQL *sql.DB, connRedis *redis.Client, logger usecase.Logger) *PostController {
 	return &PostController{
 		PostInteractor: usecase.PostInteractor{
 			AdminRepository: &AdminRepository{
